SceneDataset refactored to make PQ modifications easier.
Moved the nested functions out of SceneDatset.Open and made them
methods instead. The large nested functions made the code difficult
to read when not using the Eclipse IDE (which can collapse functions).

Split out into methods finding the root dataset and loading the bands.
This shortens and clarifies the main algorithm of Open. These are the
methods that need to be different for PQ and FC datasets.

modified:   ULA3/dataset/_scene_dataset.py